+++ tests/stubs/cycle/C.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace PhpDaTest\Stub\C;
+namespace PhpDaTest\Stub;
 
 class C
 {
     public function __construct()
     {
-        new \PhpDaTest\Stub\A\A;
+        new A;
     }
 }